Fixed signal processing and state machine

+++ project1/main.cpp
@@ -39,7 +39,7 @@ static void sigHandlerStateMachine(int sign){
 */
 static void sigHandlerMQQT(int sign){
 
-  cout <<  "State machine: Detected signal " << sign  << endl;
+  cout <<  "MQQT: Detected signal " << sign  << endl;
   if(sign == SIGINT) {
     clMqqt->stop();
   }
@@ -53,10 +53,9 @@ static void sigHandlerMQQT(int sign){
 * Singnal handler for parent
 */
 static void sigHandlerParent(int sign){
+ cout <<  "Parent: Detected signal " << sign  << endl;
 
   if (sign == SIGINT || sign == SIGUSR1) {
-
-    cout <<  "Parent: Detected signal " << sign  << endl;
     if(stmPid)
       kill(stmPid, sign);
     if(mqqtPid)  
@@ -70,21 +69,24 @@ static void sigHandlerParent(int sign){
 int main (int argc, char* argv[])
 {
   bool mqtt = false;
-  sigset_t new_set;
   stmPid = mqqtPid = 0;
-  
+  sigset_t new_set;
   cout <<  "Project1 started" << endl;
-
-  sigemptyset (&new_set);
-  sigaddset (&new_set, SIGUSR1);
-  if( sigprocmask(SIG_BLOCK, &new_set, NULL) < 0){
-      cout <<  " Could not set signal mask." << endl;
-      exit(EXIT_FAILURE);
-  }
 
   if(argc >= 2 && strcmp(argv[1], "--mqtt") == 0){
     mqtt = true;
   }
+
+/*
+  sigemptyset (&new_set);
+  sigaddset (&new_set, SIGUSR1);
+  if( sigprocmask(SIG_BLOCK, &new_set, NULL) < 0){
+    cout <<  " Could not set signal mask." << endl;
+    exit(EXIT_FAILURE);
+  }
+  else
+    cout <<  " ----  SET signal mask." << endl;
+*/
 
   switch(stmPid = fork()){
     case -1:
@@ -93,9 +95,23 @@ int main (int argc, char* argv[])
 
     case 0: //child
       {
-        if (signal(SIGINT, sigHandlerStateMachine) == SIG_ERR){
-         cout <<  "Failed set first fork handler" << endl;
-          _exit(EXIT_FAILURE);
+/*
+        sigemptyset (&new_set);
+        sigaddset (&new_set, SIGUSR1);
+        if( sigprocmask(SIG_BLOCK, &new_set, NULL) < 0){
+           cout <<  " Could not set signal mask." << endl;
+           exit(EXIT_FAILURE);
+        }
+        else
+           cout <<  " ----  SET signal mask." << endl;
+*/
+        if (signal(SIGINT, sigHandlerStateMachine) == SIG_ERR ){
+             cout <<  "Failed set first fork handler" << endl;
+             _exit(EXIT_FAILURE);
+        }
+
+        if( signal(SIGUSR1, sigHandlerStateMachine) == SIG_ERR){
+             cout <<  "Failed set first fork handler" << endl;
         }
 
         std::shared_ptr<project1::MyStateFactory> factory(new project1::MyStateFactory());
@@ -124,7 +140,8 @@ int main (int argc, char* argv[])
             break;
           case 0:
             {
-              if (signal(SIGINT, sigHandlerMQQT) == SIG_ERR){
+              if (signal(SIGINT, sigHandlerMQQT) == SIG_ERR ||
+                    signal(SIGUSR1, sigHandlerMQQT) == SIG_ERR){
                 _exit(EXIT_FAILURE);
               }
 
@@ -144,9 +161,25 @@ int main (int argc, char* argv[])
         }
       }
 
-      if (signal(SIGINT, sigHandlerParent) == SIG_ERR){
+      cout <<  "Parent task add signal SIGUSR1 " <<  stmPid << endl;
+/*
+      sigemptyset (&new_set);
+      sigaddset (&new_set, SIGUSR1);
+      if( sigprocmask(SIG_BLOCK, &new_set, NULL) < 0){
+        cout <<  " Could not set signal mask." << endl;
+        exit(EXIT_FAILURE);
+      }
+      else
+        cout <<  " ----  SET signal mask." << endl;
+*/
+      if( signal(SIGUSR1, sigHandlerParent) == SIG_ERR){
+        cout <<  "Parent handler error " <<  stmPid << endl;
+      }
+
+      if (signal(SIGINT, sigHandlerParent) == SIG_ERR ){
+        cout <<  "Parent handler error " <<  stmPid << endl;
         kill(stmPid, SIGINT);
-        
+
         if(mqtt)
           kill(mqqtPid, SIGINT);
       }
